requirements.txt
flask==3.0.3
pycryptodome==3.20.0
gunicorn==22.0.0
